refactor(referentiel): deplace le badge collectivite a droite du titre en sticky
Le badge etait centre via m-auto dans la rangee de navigation prev/next,
flottant selon la presence des liens. Extrait dans CurrentCollectiviteBadge,
il passe dans PageHeader.Actions (a droite du titre) ; la rangee redevient
purement prev/next, "Mesure suivante" tenue a droite via ml-auto.
